chatbot_ecommerce/pyproject.toml
[project]
name = "chatbot-ecommerce-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3==1.35.36",
    "chromadb==0.6.3",
    "email-validator==2.3.0",
    "fastapi==0.115.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.63",
    "langchain-openai>=0.3.27",
    "langchain-community>=0.3.5",
    "motor>=3.7.1",
    "numpy==1.26.4",
    "openai>=1.76.2",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "bcrypt==4.1.2",
    "passlib==1.7.4",
    "pdf2image==1.17.0",
    "pydantic==2.10.6",
    "pydantic-settings==2.6.0",
    "pymongo>=4.15.5",
    "pymupdf==1.24.13",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "python-pptx==1.0.2",
    "tenacity==9.0.0",
    "tiktoken==0.9.0",
    "uvicorn[standard]==0.32.0",
    "bcrypt==4.0.1",
    "streamlit>=1.54.0",
    "requests>=2.32.5",
]
# Remove the [tool.uv.workspace] section from this inner file.
